fix(tokenizer): split parser responsibilities
