--- packages/compiler/packages/tokenizer/src/parser.ts
@@ -1,982 +1,0 @@
-import type {
-	Argument,
-	AST,
-	ASTCache,
-	ASTCacheEntry,
-	BlockEndInstruction,
-	BlockEndLine,
-	BlockLine,
-	BlockPlacement,
-	BlockResultTypes,
-	BlockStartInstruction,
-	CompilerASTLine,
-	CompilerASTLines,
-	ConstantsLine,
-	ExportLine,
-	FunctionEndLine,
-	FunctionLine,
-	FunctionSignature,
-	IfEndLine,
-	IfLine,
-	ImportLine,
-	InstructionPlacement,
-	MemoryDeclarationLine,
-	ModuleLine,
-	NestedBlockPlacement,
-	PrototypeLine,
-	RegionLine,
-	SourceBlockPlacement,
-	ValidatedAST,
-} from '@8f4e/compiler-spec';
-import {
-	ArgumentType,
-	blockEndToStartInstruction,
-	blockStartInstructions,
-	compilerSourceBlockInstructionPairs,
-	DEFAULT_HOST_IMPORT_MODULE_NAME,
-	getInstructionSpec,
-	isCompilerDirectiveLine,
-	isMemoryDeclarationLine,
-	isSemanticInstructionLine,
-} from '@8f4e/compiler-spec';
-import { hashSource } from './cache';
-import instructionParser from './syntax/instructionParser';
-import isArrayDeclarationInstruction from './syntax/isArrayDeclarationInstruction';
-import isComment from './syntax/isComment';
-import isInstructionLikeLine from './syntax/isInstructionLikeLine';
-import isMemoryDeclarationInstruction from './syntax/isMemoryDeclarationInstruction';
-import { parseArgument } from './syntax/parseArgument';
-import { SyntaxErrorCode, SyntaxRulesError } from './syntax/syntaxError';
-import validateInstructionArguments from './syntax/validateInstructionArguments';
-
-/** Open-block stack frame for an `if` block while matching branches and result metadata. */
-type IfOpenBlock = {
-	instruction: 'if';
-	astIndex: number;
-	line: IfLine;
-	hasElse: boolean;
-};
-
-/** Open-block stack frame for a generic `block` while matching its end metadata. */
-type GenericOpenBlock = {
-	instruction: 'block';
-	astIndex: number;
-	line: BlockLine;
-};
-
-/** Open-block stack frame for blocks that do not need tokenizer-side line metadata. */
-type OtherOpenBlock = {
-	instruction: Exclude<BlockStartInstruction, 'if' | 'block'>;
-	astIndex: number;
-};
-
-/** Any currently open block tracked by the tokenizer's single parsing pass. */
-type OpenBlock = IfOpenBlock | GenericOpenBlock | OtherOpenBlock;
-
-/** Tracks a source block prologue so directives can be restricted to the top of a block. */
-type SourceBlockPrologue = {
-	instruction: (typeof compilerSourceBlockInstructionPairs)[number]['start'];
-	blockDepth: number;
-	isOpen: boolean;
-};
-
-/** Cached placement state maintained by the tokenizer's main loop. */
-type ParserBlockState = {
-	currentSourceBlock?: SourceBlockPlacement;
-	loopDepth: number;
-	mapDepth: number;
-	blockDepth: number;
-	ifDepth: number;
-};
-
-/** Result of validating a cached AST entry against the current source lines. */
-type ASTCacheLookupResult<TAst> = {
-	ast?: TAst;
-	hash?: number;
-};
-
-/** Accumulates module-specific lines while the tokenizer builds a validated AST. */
-type ModuleASTBuilder = {
-	type: 'module';
-	id: string;
-	moduleLine: ModuleLine;
-	regionLine?: RegionLine;
-	memoryDeclarationLines: MemoryDeclarationLine[];
-};
-
-/** Accumulates function-specific metadata while the tokenizer builds a validated AST. */
-type FunctionASTBuilder = {
-	type: 'function';
-	id: string;
-	functionLine: FunctionLine;
-	functionEndLine?: FunctionEndLine;
-	parameters: FunctionSignature['parameters'];
-	exportLine?: ExportLine;
-	exportName?: string;
-	importLine?: ImportLine;
-	import?: {
-		moduleName: string;
-		fieldName: string;
-	};
-};
-
-/** Accumulates constants-block metadata while the tokenizer builds a validated AST. */
-type ConstantsASTBuilder = {
-	type: 'constants';
-	id: string;
-	constantsLine: ConstantsLine;
-};
-
-/** Accumulates prototype memory declarations while the tokenizer builds a validated AST. */
-type PrototypeASTBuilder = {
-	type: 'prototype';
-	id: string;
-	prototypeLine: PrototypeLine;
-	memoryDeclarationLines: MemoryDeclarationLine[];
-};
-
-/** Source block builder selected from the first valid source block in parsed input. */
-type SourceBlockASTBuilder = ModuleASTBuilder | FunctionASTBuilder | ConstantsASTBuilder | PrototypeASTBuilder;
-
-/** Parsed source lines plus the source-block builder needed to materialize the final AST. */
-type ParsedCompilerSource = {
-	lines: CompilerASTLines;
-	astBuilder?: SourceBlockASTBuilder;
-};
-
-/** Physical source line after comment filtering and argument-continuation folding. */
-type SourceLine = {
-	line: string;
-	lineNumber: number;
-};
-
-/** Fast membership lookup for block-start instructions. */
-const blockStartInstructionSet = new Set<BlockStartInstruction>(blockStartInstructions);
-
-/** Fast membership lookup for top-level source block start instructions. */
-const sourceBlockStartInstructionSet: ReadonlySet<string> = new Set(
-	compilerSourceBlockInstructionPairs.map(({ start }) => start)
-);
-
-/** Fast membership lookup for top-level source block end instructions. */
-const sourceBlockEndInstructionSet: ReadonlySet<string> = new Set(
-	compilerSourceBlockInstructionPairs.map(({ end }) => end)
-);
-
-/** Narrows an instruction string to block-start instructions known by the compiler spec. */
-function isBlockStartInstruction(instruction: string): instruction is BlockStartInstruction {
-	return blockStartInstructionSet.has(instruction as BlockStartInstruction);
-}
-
-/** Narrows an instruction string to block-end instructions known by the compiler spec. */
-function isBlockEndInstruction(instruction: string): instruction is BlockEndInstruction {
-	return Object.hasOwn(blockEndToStartInstruction, instruction);
-}
-
-/** Creates the tokenizer's cached block-placement state. */
-function createParserBlockState(): ParserBlockState {
-	return {
-		loopDepth: 0,
-		mapDepth: 0,
-		blockDepth: 0,
-		ifDepth: 0,
-	};
-}
-
-/** Throws the placement syntax error with the source line metadata attached. */
-function throwInstructionNotAllowedInBlock(line: CompilerASTLine): never {
-	throw new SyntaxRulesError(SyntaxErrorCode.INSTRUCTION_NOT_ALLOWED_IN_BLOCK, undefined, {
-		lineNumber: line.lineNumber,
-		instruction: line.instruction,
-	});
-}
-
-/** Reads placement rules for regular instructions and shared memory declarations. */
-function getInstructionPlacement(line: CompilerASTLine): InstructionPlacement | undefined {
-	if (isMemoryDeclarationLine(line)) {
-		return getInstructionSpec('memoryDeclaration')?.placement;
-	}
-
-	return getInstructionSpec(line.instruction)?.placement;
-}
-
-/** Reads cached nested-block depth for a block kind without scanning the open-block stack. */
-function getNestedBlockDepth(state: ParserBlockState, block: NestedBlockPlacement): number {
-	switch (block) {
-		case 'loop':
-			return state.loopDepth;
-		case 'map':
-			return state.mapDepth;
-		case 'block':
-			return state.blockDepth;
-		case 'if':
-			return state.ifDepth;
-	}
-}
-
-/** Narrows a placement kind to source blocks that own independent ASTs. */
-function isSourceBlockPlacement(block: string): block is SourceBlockPlacement {
-	return block === 'module' || block === 'function' || block === 'constants' || block === 'prototype';
-}
-
-/** Reads block metadata for a block-start instruction from the compiler spec. */
-function getBlockStartPlacement(instruction: BlockStartInstruction): InstructionPlacement['block'] | undefined {
-	return getInstructionSpec(instruction)?.placement?.block;
-}
-
-/** Resolves the placement kind of the current immediate parent block. */
-function getOpenBlockPlacementKind(block: OpenBlock | undefined): BlockPlacement | undefined {
-	return block ? getBlockStartPlacement(block.instruction)?.kind : undefined;
-}
-
-/** Validates the immediate parent constraint for block-start instructions. */
-function validatePlacementParent(
-	line: CompilerASTLine,
-	placement: InstructionPlacement,
-	parentBlockKind: BlockPlacement | undefined
-): void {
-	const block = placement.block;
-	if (!block || block.role !== 'start' || !block.parents) {
-		return;
-	}
-
-	if (block.parents.length === 0) {
-		if (parentBlockKind !== undefined) {
-			throwInstructionNotAllowedInBlock(line);
-		}
-		return;
-	}
-
-	if (!parentBlockKind || !block.parents.includes(parentBlockKind)) {
-		throwInstructionNotAllowedInBlock(line);
-	}
-}
-
-/** Applies compiler-spec placement rules using tokenizer-maintained block state. */
-function validateInstructionPlacement(
-	line: CompilerASTLine,
-	state: ParserBlockState,
-	parentBlockKind: BlockPlacement | undefined
-): void {
-	const placement = getInstructionPlacement(line);
-	if (!placement) {
-		return;
-	}
-
-	if (state.mapDepth > 0 && placement.requiredNestedBlock !== 'map' && placement.block?.kind !== 'map') {
-		throwInstructionNotAllowedInBlock(line);
-	}
-
-	if (state.currentSourceBlock) {
-		if (placement.sourceBlocks && !placement.sourceBlocks.includes(state.currentSourceBlock)) {
-			throwInstructionNotAllowedInBlock(line);
-		}
-	} else if (!placement.topLevel && placement.block?.role !== 'start') {
-		throwInstructionNotAllowedInBlock(line);
-	}
-
-	if (placement.requiredNestedBlock && getNestedBlockDepth(state, placement.requiredNestedBlock) === 0) {
-		throwInstructionNotAllowedInBlock(line);
-	}
-
-	for (const block of placement.disallowedNestedBlocks ?? []) {
-		if (getNestedBlockDepth(state, block) > 0) {
-			throwInstructionNotAllowedInBlock(line);
-		}
-	}
-
-	validatePlacementParent(line, placement, parentBlockKind);
-}
-
-/** Updates cached block-placement state when a block starts or ends. */
-function updateParserBlockState(state: ParserBlockState, instruction: BlockStartInstruction, delta: 1 | -1): void {
-	const block = getBlockStartPlacement(instruction);
-	if (!block) {
-		return;
-	}
-
-	if (isSourceBlockPlacement(block.kind)) {
-		state.currentSourceBlock = delta === 1 ? block.kind : undefined;
-		return;
-	}
-
-	switch (block.kind) {
-		case 'loop':
-			state.loopDepth += delta;
-			return;
-		case 'map':
-			state.mapDepth += delta;
-			return;
-		case 'block':
-			state.blockDepth += delta;
-			return;
-		case 'if':
-			state.ifDepth += delta;
-			return;
-	}
-}
-
-/** Converts parsed block-end result arguments into the metadata stored on matched block lines. */
-function getResultTypesFromArguments(line: IfEndLine | BlockEndLine): BlockResultTypes {
-	return line.arguments.map(argument => argument.value as BlockResultTypes[number]);
-}
-
-/** Determines whether a memory declaration includes a default value source argument. */
-function hasExplicitMemoryDefault(instruction: string, args: Array<Argument>): boolean {
-	if (isArrayDeclarationInstruction(instruction)) {
-		return args.length > 2;
-	}
-
-	if (args.length === 0) {
-		return false;
-	}
-
-	const firstArg = args[0];
-	if (firstArg.type !== ArgumentType.IDENTIFIER) {
-		return true;
-	}
-	if (firstArg.referenceKind === 'constant') {
-		return true;
-	}
-	return args.length > 1;
-}
-
-/** Returns a cached AST only after the cheap line-count and hash checks pass. */
-function getASTCacheLookupResult<TAst>(
-	cached: ASTCacheEntry<TAst> | undefined,
-	code: string[]
-): ASTCacheLookupResult<TAst> {
-	// Optimize first compilation and obvious cache misses: only hash when an existing same-line-count entry needs validation.
-	if (!cached || cached.lineCount !== code.length) {
-		return {};
-	}
-
-	const hash = hashSource(code);
-	const cachedHash = cached.hash ?? (cached.hashInput ? hashSource(cached.hashInput.code) : undefined);
-
-	if (cachedHash === undefined) {
-		return { hash };
-	}
-
-	cached.hash = cachedHash;
-	cached.hashInput = undefined;
-	return {
-		ast: cachedHash === hash ? cached.ast : undefined,
-		hash,
-	};
-}
-
-/** Adds namespace references discovered from syntax-level arguments to a line accumulator. */
-function addReferencedNamespaceIdsFromArgument(referencedNamespaceIds: Set<string>, argument: Argument): void {
-	if (argument.type === ArgumentType.COMPILE_TIME_EXPRESSION) {
-		for (const moduleId of argument.intermoduleIds) {
-			referencedNamespaceIds.add(moduleId);
-		}
-		return;
-	}
-
-	if (argument.type !== ArgumentType.IDENTIFIER) {
-		return;
-	}
-
-	if (argument.scope === 'intermodule' && argument.targetModuleId) {
-		referencedNamespaceIds.add(argument.targetModuleId);
-	}
-}
-
-/** Creates the source-block builder represented by a parsed block-start line. */
-function createSourceBlockASTBuilder(line: CompilerASTLine): SourceBlockASTBuilder | undefined {
-	switch (line.instruction) {
-		case 'module':
-			return {
-				type: 'module',
-				id: line.arguments[0].value,
-				moduleLine: line,
-				memoryDeclarationLines: [],
-			};
-		case 'function':
-			return {
-				type: 'function',
-				id: line.arguments[0].value,
-				functionLine: line,
-				parameters: [],
-			};
-		case 'constants':
-			return {
-				type: 'constants',
-				id: line.arguments[0].value,
-				constantsLine: line,
-			};
-		case 'prototype':
-			return {
-				type: 'prototype',
-				id: line.arguments[0].value,
-				prototypeLine: line,
-				memoryDeclarationLines: [],
-			};
-		default:
-			return undefined;
-	}
-}
-
-/** Records module-only metadata while source lines are streamed into a module builder. */
-function applyModuleASTLine(builder: ModuleASTBuilder, line: CompilerASTLine): void {
-	switch (line.instruction) {
-		case '#region':
-			builder.regionLine = line;
-			return;
-		default:
-			if (!isMemoryDeclarationLine(line)) {
-				return;
-			}
-	}
-
-	builder.memoryDeclarationLines.push(line);
-}
-
-/** Records function signature, import, export, and end metadata for a function builder. */
-function applyFunctionASTLine(builder: FunctionASTBuilder, line: CompilerASTLine): void {
-	switch (line.instruction) {
-		case 'param':
-			builder.parameters.push(line.arguments[0].value as FunctionSignature['parameters'][number]);
-			return;
-		case 'functionEnd':
-			builder.functionEndLine = line;
-			return;
-		case '#export':
-			builder.exportLine = line;
-			builder.exportName = builder.exportLine.arguments[0]?.value ?? builder.id;
-			return;
-		case '#import':
-			builder.importLine = line;
-			builder.import = {
-				moduleName: DEFAULT_HOST_IMPORT_MODULE_NAME,
-				fieldName: line.arguments[0].value,
-			};
-	}
-}
-
-/** Routes a parsed line into the active source-block builder. */
-function applySourceBlockASTLine(builder: SourceBlockASTBuilder, line: CompilerASTLine): void {
-	switch (builder.type) {
-		case 'module':
-			applyModuleASTLine(builder, line);
-			return;
-		case 'function':
-			applyFunctionASTLine(builder, line);
-			return;
-		case 'constants':
-			return;
-		case 'prototype':
-			if (isMemoryDeclarationLine(line)) {
-				builder.memoryDeclarationLines.push(line);
-			}
-			return;
-	}
-}
-
-/** Materializes the final validated AST from accumulated source-block metadata. */
-function createASTFromBuilder(lines: CompilerASTLines, builder: SourceBlockASTBuilder): AST {
-	switch (builder.type) {
-		case 'module':
-			return {
-				type: 'module',
-				id: builder.id,
-				lines,
-				moduleLine: builder.moduleLine,
-				...(builder.regionLine ? { regionLine: builder.regionLine } : {}),
-				memoryDeclarationLines: builder.memoryDeclarationLines,
-			};
-		case 'function':
-			if (!builder.functionEndLine) {
-				throw new SyntaxRulesError(SyntaxErrorCode.INVALID_BLOCK_STRUCTURE, 'Expected a matching functionEnd.', {
-					lineNumber: builder.functionLine.lineNumber,
-					instruction: builder.functionLine.instruction,
-				});
-			}
-
-			return {
-				type: 'function',
-				id: builder.id,
-				lines,
-				functionLine: builder.functionLine,
-				functionEndLine: builder.functionEndLine,
-				signature: {
-					parameters: builder.parameters,
-					returns: builder.functionEndLine.arguments.map(arg => arg.value as FunctionSignature['returns'][number]),
-				},
-				...(builder.exportLine ? { exportLine: builder.exportLine, exportName: builder.exportName } : {}),
-				...(builder.importLine ? { importLine: builder.importLine, import: builder.import } : {}),
-			};
-		case 'constants':
-			return {
-				type: 'constants',
-				id: builder.id,
-				lines,
-				constantsLine: builder.constantsLine,
-			};
-		case 'prototype':
-			return {
-				type: 'prototype',
-				id: builder.id,
-				lines,
-				prototypeLine: builder.prototypeLine,
-				memoryDeclarationLines: builder.memoryDeclarationLines,
-			};
-	}
-}
-
-/** Converts parsed source into an AST and reports missing source blocks as syntax errors. */
-function toAST(parsedSource: ParsedCompilerSource): AST {
-	const { lines, astBuilder } = parsedSource;
-	const firstLine = lines[0];
-
-	if (astBuilder) {
-		return createASTFromBuilder(lines, astBuilder);
-	}
-
-	throw new SyntaxRulesError(SyntaxErrorCode.INVALID_BLOCK_STRUCTURE, 'Expected a compiler source block.', {
-		lineNumber: firstLine?.lineNumber ?? 0,
-		instruction: firstLine?.instruction,
-	});
-}
-
-/**
- * Tokenizes an instruction line, treating quoted strings as single tokens.
- * Stops at an unquoted semicolon (comment delimiter).
- * Escape sequences inside quotes are preserved for parseArgument to decode.
- */
-function tokenizeInstruction(line: string): string[] {
-	const tokens: string[] = [];
-	let i = 0;
-	const len = line.length;
-
-	while (i < len) {
-		if (/\s/.test(line[i])) {
-			i++;
-			continue;
-		}
-
-		if (line[i] === ';') break;
-
-		if (line[i] === '"') {
-			let token = '"';
-			i++;
-			while (i < len && line[i] !== '"') {
-				if (line[i] === '\\' && i + 1 < len) {
-					token += line[i] + line[i + 1];
-					i += 2;
-				} else {
-					token += line[i++];
-				}
-			}
-			if (i >= len) {
-				throw new SyntaxRulesError(SyntaxErrorCode.INVALID_STRING_LITERAL, 'Unterminated string literal');
-			}
-			token += '"';
-			i++;
-			tokens.push(token);
-		} else {
-			let token = '';
-			while (i < len && !/\s/.test(line[i]) && line[i] !== ';') {
-				token += line[i++];
-			}
-			if (token) tokens.push(token);
-		}
-	}
-
-	return tokens;
-}
-
-/** Recreates a syntax error with source line metadata from the parser context. */
-function withSyntaxLine(error: SyntaxRulesError, lineNumber: number, instruction?: string): SyntaxRulesError {
-	return new SyntaxRulesError(error.code, error.message, {
-		lineNumber,
-		instruction,
-	});
-}
-
-/** Tokenizes a physical line and attaches line metadata to tokenizer-level syntax errors. */
-function parseInstructionTokens(line: string, lineNumber: number): string[] {
-	try {
-		return tokenizeInstruction(line);
-	} catch (error) {
-		if (error instanceof SyntaxRulesError) {
-			throw withSyntaxLine(error, lineNumber);
-		}
-		throw error;
-	}
-}
-
-/** Detects physical lines that append one argument to the preceding instruction. */
-function isArgumentContinuationCandidate(line: string): boolean {
-	return /^\s*-(?=\s|;|$)/.test(line);
-}
-
-/** Removes comments and folds argument-continuation lines into logical source lines. */
-function foldArgumentContinuationLines(code: string[]): SourceLine[] {
-	const sourceLines: SourceLine[] = [];
-	let previousSourceLine: SourceLine | undefined;
-
-	for (const [lineNumber, line] of code.map((sourceLine, index) => [index, sourceLine] as const)) {
-		if (isComment(line) || !isInstructionLikeLine(line)) {
-			continue;
-		}
-
-		if (!isArgumentContinuationCandidate(line)) {
-			const sourceLine = {
-				line,
-				lineNumber,
-			};
-			sourceLines.push(sourceLine);
-			previousSourceLine = sourceLine;
-			continue;
-		}
-
-		const tokens = parseInstructionTokens(line, lineNumber);
-		const instruction = '-';
-
-		if (!previousSourceLine) {
-			throw new SyntaxRulesError(SyntaxErrorCode.INVALID_ARGUMENT, 'Argument continuation has no instruction.', {
-				lineNumber,
-				instruction,
-			});
-		}
-
-		if (tokens.length < 2) {
-			throw new SyntaxRulesError(SyntaxErrorCode.MISSING_ARGUMENT, 'Missing argument for continuation.', {
-				lineNumber,
-				instruction,
-			});
-		}
-
-		if (tokens.length > 2) {
-			throw new SyntaxRulesError(
-				SyntaxErrorCode.INVALID_ARGUMENT,
-				'Argument continuation accepts exactly one argument.',
-				{
-					lineNumber,
-					instruction,
-				}
-			);
-		}
-
-		previousSourceLine.line = `${previousSourceLine.line} ${tokens[1]}`;
-	}
-
-	return sourceLines;
-}
-
-/** Parses and syntax-validates one logical source line into an AST line. */
-export function parseLine(line: string, lineNumber: number): CompilerASTLine {
-	let instruction: string | undefined;
-	try {
-		const tokens = tokenizeInstruction(line);
-		const [first = '', ...args] = tokens;
-		instruction = first;
-		const isMemoryDeclaration = isMemoryDeclarationInstruction(instruction);
-		const shouldRecordNamespaceReferences = isMemoryDeclaration || instruction === 'const' || instruction === 'use';
-		const parsedArguments: Argument[] = [];
-		const referencedNamespaceIds = new Set<string>();
-		for (const arg of args) {
-			const parsedArgument = parseArgument(arg);
-			parsedArguments.push(parsedArgument);
-			if (shouldRecordNamespaceReferences) {
-				addReferencedNamespaceIdsFromArgument(referencedNamespaceIds, parsedArgument);
-			}
-		}
-		validateInstructionArguments(instruction, parsedArguments);
-		const [useArgument] = parsedArguments;
-		if (shouldRecordNamespaceReferences && instruction === 'use' && useArgument?.type === ArgumentType.IDENTIFIER) {
-			referencedNamespaceIds.add(useArgument.value);
-		}
-		const parsedLine = {
-			lineNumber,
-			instruction,
-			arguments: parsedArguments,
-			...(referencedNamespaceIds.size > 0 ? { referencedNamespaceIds: [...referencedNamespaceIds] } : {}),
-		} as CompilerASTLine;
-
-		if (isMemoryDeclarationLine(parsedLine)) {
-			const memoryLine: MemoryDeclarationLine = parsedLine;
-			memoryLine.hasExplicitMemoryDefault = hasExplicitMemoryDefault(instruction, parsedArguments);
-		}
-
-		return parsedLine;
-	} catch (error) {
-		if (error instanceof SyntaxRulesError) {
-			throw withSyntaxLine(
-				error,
-				lineNumber,
-				// instruction is undefined only if tokenizeInstruction threw before assignment
-				instruction ?? undefined
-			);
-		}
-		throw error;
-	}
-}
-
-/** Parses source into AST lines while validating block structure and placement in one pass. */
-function parseCompilerSource(code: string[]): ParsedCompilerSource {
-	const ast: CompilerASTLines = [];
-	let astBuilder: SourceBlockASTBuilder | undefined;
-	let onlySemanticLinesBeforeSourceBlock = true;
-	const blockStack: OpenBlock[] = [];
-	const parserBlockState = createParserBlockState();
-	const sourceBlockPrologueStack: SourceBlockPrologue[] = [];
-	const sourceLines = foldArgumentContinuationLines(code);
-
-	for (const { line, lineNumber } of sourceLines) {
-		const parsedLine = parseLine(line, lineNumber);
-		const astIndex = ast.length;
-		const isFirstASTLine = ast.length === 0;
-		const currentSourceBlockPrologue = sourceBlockPrologueStack[sourceBlockPrologueStack.length - 1];
-		const isCompilerDirective = isCompilerDirectiveLine(parsedLine);
-		const isInOpenSourceBlockPrologue =
-			currentSourceBlockPrologue?.isOpen && currentSourceBlockPrologue.blockDepth === blockStack.length;
-
-		if (isCompilerDirective && isInOpenSourceBlockPrologue) {
-			parsedLine.isBlockPrologue = true;
-		} else if (isCompilerDirective) {
-			throw new SyntaxRulesError(SyntaxErrorCode.COMPILER_DIRECTIVE_MUST_BE_PROLOGUE, undefined, {
-				lineNumber,
-				instruction: parsedLine.instruction,
-			});
-		}
-
-		if (isInOpenSourceBlockPrologue && !isCompilerDirective) {
-			currentSourceBlockPrologue.isOpen = false;
-		}
-
-		validateInstructionPlacement(
-			parsedLine,
-			parserBlockState,
-			getOpenBlockPlacementKind(blockStack[blockStack.length - 1])
-		);
-
-		ast.push(parsedLine);
-		if (!astBuilder) {
-			const candidateBuilder = createSourceBlockASTBuilder(parsedLine);
-			if (
-				candidateBuilder &&
-				(isFirstASTLine || (candidateBuilder.type === 'function' && onlySemanticLinesBeforeSourceBlock))
-			) {
-				astBuilder = candidateBuilder;
-			}
-		}
-		if (astBuilder) {
-			applySourceBlockASTLine(astBuilder, parsedLine);
-		} else if (!isSemanticInstructionLine(parsedLine)) {
-			onlySemanticLinesBeforeSourceBlock = false;
-		}
-
-		if (isBlockStartInstruction(parsedLine.instruction)) {
-			switch (parsedLine.instruction) {
-				case 'if':
-					blockStack.push({
-						instruction: 'if',
-						astIndex,
-						line: parsedLine,
-						hasElse: false,
-					});
-					break;
-				case 'block':
-					blockStack.push({
-						instruction: 'block',
-						astIndex,
-						line: parsedLine,
-					});
-					break;
-				default:
-					blockStack.push({
-						instruction: parsedLine.instruction,
-						astIndex,
-					});
-			}
-			updateParserBlockState(parserBlockState, parsedLine.instruction, 1);
-
-			if (sourceBlockStartInstructionSet.has(parsedLine.instruction)) {
-				sourceBlockPrologueStack.push({
-					instruction: parsedLine.instruction as SourceBlockPrologue['instruction'],
-					blockDepth: blockStack.length,
-					isOpen: true,
-				});
-			}
-			continue;
-		}
-
-		if (parsedLine.instruction === 'else') {
-			const openBlock = blockStack[blockStack.length - 1];
-
-			if (!openBlock || openBlock.instruction !== 'if') {
-				throw new SyntaxRulesError(
-					SyntaxErrorCode.INVALID_BLOCK_STRUCTURE,
-					'Unexpected else without a matching open if block.',
-					{
-						lineNumber,
-						instruction: parsedLine.instruction,
-					}
-				);
-			}
-
-			if (openBlock.hasElse) {
-				throw new SyntaxRulesError(
-					SyntaxErrorCode.INVALID_BLOCK_STRUCTURE,
-					'Unexpected else: if blocks may only contain one else branch.',
-					{
-						lineNumber,
-						instruction: parsedLine.instruction,
-					}
-				);
-			}
-
-			openBlock.hasElse = true;
-			continue;
-		}
-
-		if (!isBlockEndInstruction(parsedLine.instruction)) {
-			continue;
-		}
-
-		const endInstruction = parsedLine.instruction;
-		const expectedStartInstruction = blockEndToStartInstruction[endInstruction];
-		const openBlock = blockStack.pop();
-
-		if (!openBlock) {
-			throw new SyntaxRulesError(
-				SyntaxErrorCode.INVALID_BLOCK_STRUCTURE,
-				`Unexpected ${endInstruction} without a matching open ${expectedStartInstruction} block.`,
-				{
-					lineNumber,
-					instruction: parsedLine.instruction,
-				}
-			);
-		}
-
-		if (openBlock.instruction !== expectedStartInstruction) {
-			throw new SyntaxRulesError(
-				SyntaxErrorCode.INVALID_BLOCK_STRUCTURE,
-				`Unexpected ${endInstruction}: expected ${openBlock.instruction} to be closed before ending ${expectedStartInstruction}.`,
-				{
-					lineNumber,
-					instruction: parsedLine.instruction,
-				}
-			);
-		}
-		updateParserBlockState(parserBlockState, openBlock.instruction, -1);
-
-		if (sourceBlockEndInstructionSet.has(endInstruction)) {
-			sourceBlockPrologueStack.pop();
-		}
-
-		if (endInstruction !== 'ifEnd' && endInstruction !== 'blockEnd') {
-			continue;
-		}
-
-		if (parsedLine.instruction === 'ifEnd' && openBlock.instruction === 'if') {
-			const resultTypes = getResultTypesFromArguments(parsedLine);
-			openBlock.line.ifBlock = {
-				matchingIfEndIndex: astIndex,
-				resultTypes,
-				hasElse: Boolean(openBlock.hasElse),
-			};
-			parsedLine.ifEndBlock = {
-				matchingIfIndex: openBlock.astIndex,
-				resultTypes,
-			};
-		} else if (parsedLine.instruction === 'blockEnd' && openBlock.instruction === 'block') {
-			const resultTypes = getResultTypesFromArguments(parsedLine);
-			openBlock.line.blockBlock = {
-				matchingBlockEndIndex: astIndex,
-				resultTypes,
-			};
-			parsedLine.blockEndBlock = {
-				matchingBlockIndex: openBlock.astIndex,
-				resultTypes,
-			};
-		}
-	}
-
-	if (blockStack.length > 0) {
-		const openBlock = blockStack[blockStack.length - 1];
-		const openLine = ast[openBlock.astIndex];
-
-		throw new SyntaxRulesError(SyntaxErrorCode.INVALID_BLOCK_STRUCTURE, `Unclosed ${openBlock.instruction} block.`, {
-			lineNumber: openLine.lineNumber,
-			instruction: openBlock.instruction,
-		});
-	}
-
-	return { lines: ast, astBuilder };
-}
-
-/** Compiles source to validated AST lines, using the optional cache for repeated parses. */
-export function compileToASTLines(
-	code: string[],
-	cache?: ASTCache<CompilerASTLines>,
-	cacheKey?: string
-): CompilerASTLines {
-	const cached = cacheKey !== undefined ? cache?.entries.get(cacheKey) : undefined;
-	const cachedLookup = getASTCacheLookupResult(cached, code);
-	if (cachedLookup.ast) {
-		cache!.stats.hits++;
-		return cachedLookup.ast;
-	}
-	if (cache && cacheKey !== undefined) {
-		cache.stats.misses++;
-	}
-
-	const ast = parseCompilerSource(code).lines;
-
-	if (cache && cacheKey !== undefined) {
-		const entry: ASTCacheEntry<CompilerASTLines> = {
-			ast,
-			hashInput: {
-				code,
-			},
-			lineCount: code.length,
-		};
-		if (cachedLookup.hash !== undefined) {
-			entry.hash = cachedLookup.hash;
-			entry.hashInput = undefined;
-		}
-		cache.entries.set(cacheKey, entry);
-	}
-
-	return ast;
-}
-
-/** Compiles source into a validated source-block AST, using the optional cache when available. */
-export function compileToAST(code: string[], cache?: ASTCache<ValidatedAST>, cacheKey?: string): ValidatedAST {
-	const cached = cacheKey !== undefined ? cache?.entries.get(cacheKey) : undefined;
-	const cachedLookup = getASTCacheLookupResult(cached, code);
-	if (cachedLookup.ast) {
-		cache!.stats.hits++;
-		return cachedLookup.ast;
-	}
-	if (cache && cacheKey !== undefined) {
-		cache.stats.misses++;
-	}
-
-	const parsedSource = parseCompilerSource(code);
-	const group = toAST(parsedSource) as ValidatedAST;
-
-	if (cache && cacheKey !== undefined) {
-		const entry: ASTCacheEntry<ValidatedAST> = {
-			ast: group,
-			hashInput: {
-				code,
-			},
-			lineCount: code.length,
-		};
-		if (cachedLookup.hash !== undefined) {
-			entry.hash = cachedLookup.hash;
-			entry.hashInput = undefined;
-		}
-		cache.entries.set(cacheKey, entry);
-	}
-
-	return group;
-}
-
-export { instructionParser };
